[REFACTOR] Move files to correct package, Refactor launch custom tabs
