+++ imx-core-sdk-android/src/main/java/com/immutable/sdk/model/Erc721Asset.kt
@@ -4,7 +4,6 @@ import com.immutable.sdk.api.model.SignableToken
 import com.immutable.sdk.api.model.Token
 import com.immutable.sdk.api.model.TokenData
 import com.immutable.sdk.utils.Constants.ERC721_AMOUNT
-import com.immutable.sdk.utils.TokenType
 
 /**
  * This is an ERC721 wrapper for the [Token] API model
