+++ imx-core-sdk-android/src/main/java/com/immutable/sdk/model/TokenType.kt
@@ -1,4 +1,4 @@
-package com.immutable.sdk.utils
+package com.immutable.sdk.model
 
 internal enum class TokenType {
     ETH, ERC20, ERC721
